encryption support added to stackedfs and overlay pdevs support
Changelog
---------
* implemented advanced persistence devices (pdevs) design - see below
* added encryption support
* changed persistence compliance configuration file layout
  * removed "type" field (currently only supporting loopback by default)
  * added "crypto" field (currently only supporting luks/none)
* modularized code to be less complex and more generic
    * re-write of persistence code in setup_unionfs
    new casper-helpers functions
    * remove_loop
    * crypto_open
    * crypto_close
    * process_personalize_conf
    * process_pdevs_conf
* code cleanup

Design overview
---------------
When casper is installed via the fab, a random and unique 2048bit
keyfile is generated and embedded in the initramfs as /etc/pdevs.key.
The pdevs keyfile will allow casper to access the encrypted luks
loopback device called "pdevs" which contains the following:
	* pdevs.${pcom}.conf
	* personalize.conf
	* keyfiles for the pdevs

PCOM (persistence compliance) is a variable hardcoded into
/etc/casper.conf which differenciates between pdevs created for
different releases. This allows the user to have multiple pdev's on
his system for different releases, without the worry of different
versions not complying with one another and therefore breaking
something...

The pdevs loopback must be created by the user prior to booting in
persistent mode. Currently, the user will have to add the keyfile
casper has access to, to the pdevs luks device, this issue will be
extended in upcoming commits.

We currently only support luks encryption via loopback, and casper
unlocking using keyfiles, more support to come. The pdevs loopback
must be encrypted (in this commit), but the actual persistent devices
don't have to be.

Example pdevs.${pcom}.conf (see /etc/casper.conf for ${pcom})

    # Persistence Compliance configuration file
    #
    #   overlay [true | false]
    #   crypto  [luks | none]
    #
    # label             mountp     overlay     crypto
    # -----             ------     -------     ------
    ${pcom}.root.pdev   /          true        luks
    ${pcom}.home.pdev   /home      false       luks

The pdev's will be stacked in the order they are specified. The lowest
layer in the stack is the squashfs rootfs, with a tmpfs overlay ontop
of it. If a root "/" persistent overlay is specified, it will take
presidence.

It is recommended that the pdev labels (filename) include the PCOM,
but it is not mandatory. The pdev's can be stored on any partition
with one of the following filesystems: vfat, ext2, ext3.

The persistent devices key files must be stored in the pdevs encrypted
loopback, and must be called ${label}.key

The personalize configuration file needed sets will create and set the
username, fullname and hostname. It must be present on first boot into
persistent mode if using a root "/" persistent overlay - otherwise the
default user "sterile" (uid 999) and hostname "sterile" will be set,
and saved in the overlay.

Example personalize.conf

    # Personalize configuration file
    export USERNAME="loni"
    export USERFULLNAME="persistent session user"
    export HOST="thought"