interactive support through usplash (unlock + add embedded key)
casper will communicate with the user either via the console, or via
usplash (if it is running), inorder to unlock pdev, and approve addition
of embedded key.

note: i am using an ugly workaround due to the version of usplash we are
using (edgy release) is buggy. when we upgrade it, i will fix the code
accordingly.