eliminated personalization
"Personalization" should be eliminated entirely. Its an annoying, retarded
misfeature that interrupts the boot sequence with slow to react, nearly
invisible prompts. I understand this is sort of a necessary evil right now with
ectoplasm passphrase, but the benefit of having a "personalized" username and
hostname doesn't even come close to justifying further prompts.

I suspect the only reason you don't mind that much is that you never reboot. It
pisses me off even in vmware.

Besides, on a single user system you don't *need* to have a personalized
username or hostname anyway.